Municipality info (legacy)

A dataset containing the municipality codes. The database is outdated, as it does not consider all municipalities.

cadmun

Format

a data frame with 5657 rows e 35 variables:

information about border, Amazon, capital, longitude, latitude, among others

Source

cadmun

Examples

summary(cadmun)
#>      MUNCOD          MUNCODDV         SITUACAO            MUNSINP     
#>  Min.   :110000   Min.   :1100000   Length:5657        Min.   : 2000  
#>  1st Qu.:251276   1st Qu.:2512762   Class :character   1st Qu.:10013  
#>  Median :314870   Median :3148707   Mode  :character   Median :14040  
#>  Mean   :327254   Mean   :3272545                      Mean   :14032  
#>  3rd Qu.:412230   3rd Qu.:4122305                      3rd Qu.:19370  
#>  Max.   :530010   Max.   :5300108                      Max.   :28141  
#>                                                        NA's   :69     
#>     MUNSIAFI      MUNNOME            MUNNOMEX            OBSERV         
#>  Min.   :   1   Length:5657        Length:5657        Length:5657       
#>  1st Qu.:1614   Class :character   Class :character   Class :character  
#>  Median :4392   Mode  :character   Mode  :character   Mode  :character  
#>  Mean   :4529                                                           
#>  3rd Qu.:7190                                                           
#>  Max.   :9997                                                           
#>  NA's   :77                                                             
#>    MUNSINON          MUNSINONDV          AMAZONIA          FRONTEIRA        
#>  Length:5657        Length:5657        Length:5657        Length:5657       
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>    CAPITAL              UFCOD          MESOCOD        MICROCOD    
#>  Length:5657        Min.   :11.00   Min.   :1100   Min.   :11000  
#>  Class :character   1st Qu.:25.00   1st Qu.:2503   1st Qu.:25016  
#>  Mode  :character   Median :31.00   Median :3110   Median :31049  
#>                     Mean   :32.57   Mean   :3262   Mean   :32596  
#>                     3rd Qu.:41.00   3rd Qu.:4107   3rd Qu.:41027  
#>                     Max.   :53.00   Max.   :5301   Max.   :53001  
#>                                                                   
#>     MSAUDCOD       RSAUDCOD       CSAUDCOD         CIRCOD         RMETRCOD   
#>  Min.   :1100   Min.   :1100   Min.   :11000   Min.   :11000   Min.   :1100  
#>  1st Qu.:2590   1st Qu.:2507   1st Qu.:25900   1st Qu.:25011   1st Qu.:2590  
#>  Median :3107   Median :3117   Median :31049   Median :31055   Median :3190  
#>  Mean   :3292   Mean   :3267   Mean   :33118   Mean   :32595   Mean   :3335  
#>  3rd Qu.:4190   3rd Qu.:4117   3rd Qu.:41900   3rd Qu.:41017   3rd Qu.:4190  
#>  Max.   :5301   Max.   :5301   Max.   :53001   Max.   :53001   Max.   :5301  
#>                                                                              
#>      AGLCOD        ANOINST         ANOEXT        SUCESSOR      
#>  Min.   :1100   Min.   :1534   Min.   :1988   Min.   :1700400  
#>  1st Qu.:2590   1st Qu.:1935   1st Qu.:1988   1st Qu.:1705607  
#>  Median :3190   Median :1960   Median :1988   Median :1711100  
#>  Mean   :3340   Mean   :1945   Mean   :1988   Mean   :1753633  
#>  3rd Qu.:4190   3rd Qu.:1989   3rd Qu.:1988   3rd Qu.:1717008  
#>  Max.   :5301   Max.   :2013   Max.   :2002   Max.   :4302105  
#>                 NA's   :92     NA's   :5596   NA's   :5596     
#>     LATITUDE          LONGITUDE          ALTITUDE          AREA          
#>  Min.   :-281.615   Min.   :-486.70   Min.   :   0.0   Length:5657       
#>  1st Qu.: -22.786   1st Qu.: -50.80   1st Qu.: 138.0   Class :character  
#>  Median : -17.732   Median : -46.36   Median : 391.0   Mode  :character  
#>  Mean   : -16.296   Mean   : -45.64   Mean   : 401.6                     
#>  3rd Qu.:  -8.087   3rd Qu.: -41.08   3rd Qu.: 625.0                     
#>  Max.   :   4.596   Max.   :   0.00   Max.   :1628.0                     
#>                                                                          
#>       lat                lon               cod           regiao         
#>  Min.   :-281.615   Min.   :-486.70   Min.   :11.00   Length:5657       
#>  1st Qu.: -22.786   1st Qu.: -50.80   1st Qu.:25.00   Class :character  
#>  Median : -17.732   Median : -46.36   Median :31.00   Mode  :character  
#>  Mean   : -16.296   Mean   : -45.64   Mean   :32.57                     
#>  3rd Qu.:  -8.087   3rd Qu.: -41.08   3rd Qu.:41.00                     
#>  Max.   :   4.596   Max.   :   0.00   Max.   :53.00                     
#>                                                                         
#>       uf            municipio_uf      
#>  Length:5657        Length:5657       
#>  Class :character   Class :character  
#>  Mode  :character   Mode  :character  
#>                                       
#>                                       
#>                                       
#>