Municipality info

A dataset containing useful data for joining with city datasets, such as city codes, acronyms of the federative units and regions.

muni

Format

a data frame with 5572 rows and 17 columns:

muni_id

IBGE code (7 digits)

muni_id_6

IBGE code (6 digits)

muni_nm

Municipality name (original IBGE)

muni_nm_clean

Clean municipality name (without accents, upper case)

uf_nm

Name of the original federative unit

uf_sigla

Federative unit initials

uf_id

code of IBGE federative unit

regiao_nm

region name

tse_id

municipality TSE code

rf_id

Brazilian Federal Revenue code of the municipality

bcb_id

Central Bank of Brazil code of the municipality

existia_1991

1 if municipality existed in 1991, 0 otherwise

existia_2000

1 if municipality existed in 2000, 0 otherwise

existia_2010

1 if municipality existed in 2010, 0 otherwise

lon

longitude based on shapefile centroid

lat

latitude based on shapefile centroid

capital

TRUE if federative unit is a capital, FALSE otherwise

Source

Municipalities dataset enriched with external data.

Examples

summary(muni)
#>    muni_id           muni_id_6           muni_nm          muni_nm_clean     
#>  Length:5572        Length:5572        Length:5572        Length:5572       
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>     uf_nm             uf_sigla            uf_id            regiao_nm        
#>  Length:5572        Length:5572        Length:5572        Length:5572       
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>     tse_id             rf_id              bcb_id           existia_1991   
#>  Length:5572        Length:5572        Length:5572        Min.   :0.0000  
#>  Class :character   Class :character   Class :character   1st Qu.:1.0000  
#>  Mode  :character   Mode  :character   Mode  :character   Median :1.0000  
#>                                                           Mean   :0.8063  
#>                                                           3rd Qu.:1.0000  
#>                                                           Max.   :1.0000  
#>                                                           NA's   :2       
#>   existia_2000     existia_2010         lon              lat         
#>  Min.   :0.0000   Min.   :0.0000   Min.   :-73.44   Min.   :-33.653  
#>  1st Qu.:1.0000   1st Qu.:1.0000   1st Qu.:-50.91   1st Qu.:-22.831  
#>  Median :1.0000   Median :1.0000   Median :-46.52   Median :-18.113  
#>  Mean   :0.9887   Mean   :0.9991   Mean   :-46.25   Mean   :-16.458  
#>  3rd Qu.:1.0000   3rd Qu.:1.0000   3rd Qu.:-41.40   3rd Qu.: -8.503  
#>  Max.   :1.0000   Max.   :1.0000   Max.   :-32.42   Max.   :  4.685  
#>  NA's   :2        NA's   :2                                          
#>   capital       
#>  Mode :logical  
#>  FALSE:5545     
#>  TRUE :27       
#>                 
#>                 
#>                 
#>